backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
requests>=2.31.0
pandas>=2.2.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
